Create ideal3.hs
I finally learned how to make the Haskell versions compilable. I didn't add in commandline arguments for initial parameters (yet).
The proclivity is basically the same as in the Java version. And depending on the initial conditions, the output is exactly the same.
The simple version is not implemented, due to seeming more difficult.