Update ideal3.hs
Modified to accommodate 4 experiments.

Changing valences or results would only be meaningful if I concoct a more crafty environment.
Additional variables may need to be introduced for those environments anyway.

The agent takes the first pleasing pattern it finds. If it finds one with only E1 and E2 it sticks with that. There is a learning curve for reaching E4. Simply taking the next experiment instead of one not anticipated worked worse.