Create ideal4a.hs
Hmm, I failed to properly implement lesson 4. I got a system that manages to solve Environment040 IF positive valences are set to 2, not 1. At first it learned increasingly long sequences (A works, so it learns AA and AAA and AAAA...), but upon changing how it calculates the weight so that zero valence doesn't mean zero proclivity, it worked properly. It also works if I set valence to 2.5 (so that the proper sequence adds up to something positive). This may be because I hacked around part of the abstraction by comparing all interactions that recommend the same sequence of experiments. It learns the correct sequence by step 7 though, which is pretty cool ^_^.

Ultimately, the experiment-based abstraction that takes a failed interaction <e r> and creates a new interaction <e r1> is useful for lowering failed interactions that 'seem' good (positive valence). (And that may be part of why it doesn't work on Environment41.)
I'm not quite sure how to incorporate that into this model.

Maybe all will become clear later. Anyway, that's enough on this for now. I've learned a lot =D